Refactor: Storing vertices and their triangle indices for the convex shape instead of 3 vertices for each triangle Control the step of the algorithm through GUI Initial points is know marked with their indices

+++ src/doubly_linked_list.h
@@ -20,7 +20,7 @@ DNode* DListNewNode(void* data);
 void DListPushBack(DoublyLinkedList* list, void* data);
 void DListRemoveNode(DoublyLinkedList *list, DNode* node);
 int DListIsEmpty(const DoublyLinkedList* list);
-void *DListToArray(DoublyLinkedList *list, size_t elementSize, size_t *outSize);
+void *DListToArray(DoublyLinkedList *list, size_t elementSize, int *outSize);
 void DListClear(DoublyLinkedList *list);
 
 #endif
